Make missing key configuration a compile-time error
Throwing exceptions is not idiomatic in Scala land. Try is better, but if we
wrapped every API call in a Try because of the possibility of a missing API
key, then we’d be giving callers an awkward Try[Future[Response]] to unpack for
every method, which is not much less onerous than forcing them to handle with
try/catch.

This factors client functionality into a separate trait for every access level
(read, write, master). You must mix one of these traits into a Client instance
in order to be able to make calls requiring that level of access, and each
level’s trait defines its key field as abstract so that failing to provide it
*errors at compile time*. No more runtime checks.

If you want to read and write, for example, you initialize a client like this:

    val keen = Client.new("projectID") with Reader with Writer {
      val readKey = "myReadKey"
      val writeKey = "myWriteKey"
    }

As a next step, we'll have the keys default to values from a config file so
that this will become simpler:

    val keen = Client.new("projectID") with Reader with Writer

The Master trait allows reading, writing, and master operations with only one
mixin and key configuration value required, which of course should be
discouraged when master isn't needed, but it reflects semantically how the API
handles master authorization.

This needs documentation updates and new unit tests, but since it's still an
incremental step with a config file coming next (via Typesafe Config), I wanted
to see how well that works out before fully updating everything.