Further restrict AccessLevel’s self-type to Client
In ongoing work to make Client configuration less dependent on lots of
constructor params, it was becoming clear that the AccessLevels would
want to use a few more of Client’s facilities beyond just the HTTP
adapter.

We really intend the usage of these to be mixing into Client, and you
can say that a Writer "is a" (augmented) Client, so let’s express that
clearly and conservatively.