Use better types for request batching config settings
We do gymnastics in code to treat Longs as time durations -- Typesafe
Config can do better than that and it is nicer both internally and for
library users. It's sometimes less obvious when we're use Java threading
APIs, but if we refactor to use more modern Scala concurrency primitives
it will stand out further.

This changes config values that represent time durations to use
scala.concurrent.duration.FiniteDuration as appropriate.

It also moves default values into reference.conf and out of code where
they are more opaque and more likely to suffer copy-and-paste bugs and
repetitious `getOrElse` proliferation. Keys are moved out of optional
because we provide default values for them. Leaving settings completely
optional is rife with pain and Typesafe Config shuns it unless it is
truly warranted, like our read/write/master key settings. See #20 and:

https://github.com/typesafehub/config#how-to-handle-defaults