Move batched write functionality to a Client subclass
I raised an eyebrow when the shutdown() method added to Writer to
terminate threadpools could not call super() to invoke Client’s
shutdown(), because Writer is a trait with a self-type, not a
subclass. This lead me to think further:

Writer does not have an “is a” relationship to Client. Multiple
AccessLevels like Writer can be mixed into a Client, and the
AccessLevel abstraction is about *what* API calls a Client can
make, not *how* it makes them (e.g. with delayed batching).

Thus I felt that batched write functionality really belongs in a
Client subclass instead. This choice seems supported by the facts
that the natural super() call is now possible, “constants” for
flush intervals can now truly be constant (not allocated per Client
instance) by being defined in a companion object, there is no
chance of inadvertently creating a needless threadpool when using
Writer with no desire for batched writes, etc.