Don’t run Client-only specs for BatchWriterClient, for now
I had hastily extended ClientSpec in the BatchWriterClientSpec to retain
access to the fakes like OkHttpAdapter, but that ran a bunch of specs
that use a standard Client.

These suites need some refactoring to share common base tests with an
injected custom Client instance/subclass. That's a noisier change so I'm
going to save it for a bit later.