Nix unused/incomplete attempt counting
Vestiges of the Java client’s attempt counting were brought over when
adding batched write queueing in #38, but this functionality was not
integrated at all in the Scala client and configuration yet.

Until it’s ready, with configuration and documentation, let’s axe it.
It’s easier to resurrect this than bear the maintenance costs of dead
code IMHO.

For some comparison of what was missing, see keenlabs/KeenClient-Java#29